+++ runtime/kyo/src/main/scala/mongo4s/kyo/KyoBridgeInstance.scala
@@ -28,7 +28,7 @@ trait KyoBridgeInstance:
       withTimeout(
         Sync.defer(
           Abort.catching(
-            Async.fromCompletableFuture(PublisherCollector.collect(publisher, limit))
+            Async.fromCompletableFuture(PublisherCollector.collect(RsBridgeSupport.translating(publisher), limit))
           )
         )
       )
@@ -47,7 +47,7 @@ trait KyoBridgeInstance:
       withTimeout(
         Sync.defer(
           Abort.catching(
-            Async.fromCompletableFuture(PublisherCollector.drain(publisher))
+            Async.fromCompletableFuture(PublisherCollector.drain(RsBridgeSupport.translating(publisher)))
           )
         )
       )
@@ -65,7 +65,7 @@ trait KyoBridgeInstance:
       given Tag[Emit[Chunk[A]]] = streamable.emitTag
       given Tag[Poll[Chunk[A]]] = streamable.pollTag
 
-      Stream.unwrap(fromPublisher(publisher, config.bufferSize)).handle(Scope.run)
+      Stream.unwrap(fromPublisher(RsBridgeSupport.translating(publisher), config.bufferSize)).handle(Scope.run)
     end stream
 
 object KyoBridgeInstance extends KyoBridgeInstance: